Improvements due to GitHub comments
Entries in the writer history cache will only be pushed to the end
of the expiration queue if they have not already expired, this is
to preserve the ordering of the queue even when an instance has a
queued deadline which has already passed
Improved unittest for the supplied scenario

+++ src/core/ddsi/include/dds/ddsi/ddsi_deadline.h
@@ -36,7 +36,7 @@ struct deadline_elem {
   struct ddsrt_circlist_elem e;
   ddsrt_mtime_t t_deadline;
   ddsrt_mtime_t t_last_trigger;
-  uint32_t deadlines_missed;
+  uint64_t deadlines_missed;
 };
 
 /** @component deadline_qos */
